+++ src/interpreter/builtins/iterable.rs
@@ -33,6 +33,31 @@ fn map<'a>(
     .boxed()
 }
 
+fn filter<'a>(
+    env: &'a mut Env,
+    receiver: &'a Value,
+    args: &'a [Value],
+) -> BoxFuture<'a, Result<Value>> {
+    async move {
+        let mut values = vec![];
+        for v in receiver.get_items()? {
+            match args.first() {
+                Some(Value::Func(func)) => match func.execute(env, &[v.clone()]).await? {
+                    Value::Bool(true) => values.push(v),
+                    Value::Bool(false) => continue,
+                    _ => bail!("filter function must return a boolean"),
+                },
+                _ => bail!("filter function expects a function as an argument"),
+            };
+        }
+        match receiver.get_type() {
+            Type::Array(t) => Ok(Value::Array(values, t.clone())),
+            ty => bail!("cannot map to type {}", ty),
+        }
+    }
+    .boxed()
+}
+
 pub fn iter_len(_env: &Env, arg: &Value) -> Result<Value> {
     arg.len().map(Into::into)
 }
@@ -42,6 +67,11 @@ lazy_static! {
         "map",
         map,
         vec![vec![FunctionParam::new("f", Type::Function)]]
+    );
+    pub static ref ITER_FILTER: Arc<dyn FunctionDef> = AsyncMethod::arc(
+        "filter",
+        filter,
+        vec![vec![FunctionParam::new("p", Type::Function)]]
     );
     pub static ref ITER_LEN: Arc<dyn FunctionDef> = SyncProperty::arc("length", iter_len);
 }